refactor: 유틸함수 분리, 필요없는 컴포넌트 삭제
