+++ week3/assignment/src/App.jsx
@@ -1,13 +1,13 @@
 import { Global, ThemeProvider } from "@emotion/react";
 import GlobalStyle from "./styles/global";
 import theme from "./styles/theme";
-import Content from "./components/Content";
+import Game from "./pages/Game";
 
 function App() {
   return (
     <ThemeProvider theme={theme}>
       <Global styles={GlobalStyle} />
-      <Content />
+      <Game />
     </ThemeProvider>
   );
 }
