+++ week3/assignment/src/pages/Game.jsx
@@ -1,49 +1,27 @@
-import Header from "./Header";
-import NumberGame from "./NumberGame";
-import styled from "@emotion/styled";
-import Ranking from "./Ranking";
 import { useState } from "react";
 import useTimer from "../hooks/useTimer";
-import { formatTime } from "../utils/formatTime";
+import { formatTime, formatDate } from "../utils/format";
+import styled from "@emotion/styled";
+import Header from "../components/Header";
+import NumberGame from "../components/NumberGame";
+import Ranking from "../components/Ranking";
 
-const Content = () => {
+const Game = () => {
   const [content, setContent] = useState("GAME");
-  const handleChangeContent = (content) => {
-    setContent(content);
-  };
-
-  const { time, startTimer, resetTimer } = useTimer();
-
-  if (!localStorage.getItem("ranking")) {
-    localStorage.setItem("ranking", JSON.stringify([]));
-  }
 
   const [gameRanking, setGameRanking] = useState(
     JSON.parse(localStorage.getItem("ranking"))
   );
 
+  const { time, startTimer, resetTimer } = useTimer();
+
+  const handleChangeContent = (content) => {
+    setContent(content);
+  };
+
   const handleSetLocalStorage = () => {
-    const today = new Date();
-
-    // 시간 형식화
-    const hours = today.getHours();
-    const minutes = today.getMinutes();
-    const seconds = today.getSeconds();
-
-    // 오전/오후 표시
-    const ampm = hours >= 12 ? "오후" : "오전";
-    const formattedHours = hours % 12 === 0 ? 12 : hours % 12; // 12시간제 변환
-
-    // 최종 날짜 및 시간 포맷팅
-    const formattedDate = `${today.getFullYear()}. ${
-      today.getMonth() + 1
-    }. ${today.getDate()}. ${ampm} ${formattedHours}:${String(minutes).padStart(
-      2,
-      "0"
-    )}:${String(seconds).padStart(2, "0")}`;
-
     const newRank = {
-      time: `${formattedDate}`,
+      time: `${formatDate()}`,
       level: "level 01",
       play: `${formatTime(time)}`,
     };
@@ -57,6 +35,10 @@ const Content = () => {
     setGameRanking([]);
     localStorage.setItem("ranking", []);
   };
+
+  if (!localStorage.getItem("ranking")) {
+    localStorage.setItem("ranking", JSON.stringify([]));
+  }
 
   return (
     <ContentLayout>
@@ -82,7 +64,7 @@ const Content = () => {
   );
 };
 
-export default Content;
+export default Game;
 
 const ContentLayout = styled.div`
   display: flex;
