+++ week3/assignment/src/utils/createRandomArray.js
@@ -1,0 +1,5 @@
+export const createRandomArray = (start, end) => {
+  return Array.from({ length: end - start + 1 }, (_, i) => start + i).sort(
+    () => Math.random() - 0.5
+  );
+};
